S3 handler module with put_object (#48)
## Description
Creating a new module to house S3 related functions and adding a new
function for uploading file objects to S3.

## Related Issues
closes #39
