+++ tests/unit/test_main.py
@@ -44,26 +44,3 @@ class TestHandler:
 
         assert result["file_contents"] == []
         assert len(result["file_contents"]) == 0
-
-
-class TestCreateS3Client:
-    def test_create_s3_client(self, moto_setup):
-        s3_client = main.create_s3_client()
-        assert s3_client.meta.endpoint_url == "https://s3.amazonaws.com"
-        assert s3_client.meta.region_name == "us-east-1"
-        assert s3_client._get_credentials().secret_key == "test"
-        assert s3_client._get_credentials().access_key == "test"
-
-
-class TestGetFileContentFromS3Event:
-    def test_get_file_content_from_s3_event(self, moto_setup):
-        moto_setup.put_object(
-            Bucket=moto_setup.bucket_name, Key="test.txt", Body=b"This eICR has errors"
-        )
-
-        event = {
-            "detail": {"bucket": {"name": moto_setup.bucket_name}, "object": {"key": "test.txt"}}
-        }
-
-        content = main.get_file_content_from_s3_event(event)
-        assert content == b"This eICR has errors"
